Builded

+++ src/domain/entities/shop/department/department.h
@@ -1,22 +1,23 @@
 #pragma once
 
-#include "utils/tagged.h"
-#include "domain/value_data/types.h"
-
+#include <cstdint>
 #include <optional>
+
+#include "tagged.h"
+#include "value_data/types.h"
 
 namespace domain {
 
 namespace detail {
-    struct Department{};
-} // namespace detail
+struct Department {};
+}  // namespace detail
 
-using DepartmentId = utils::Tagged<int, detail::Department>;
+using DepartmentId = utils::Tagged<uint64_t, detail::Department>;
 using DepartmentIdHasher = utils::TaggedHasher<DepartmentId>;
 
 struct Department {
     DepartmentId department_id;
     std::string name;
 };
-    
-} // namespace domain
+
+}  // namespace domain
