+++ src/domain/entities/shop/profession/profession.h
@@ -1,9 +1,10 @@
 #pragma once
 
+#include <cstdint>
 #include <optional>
 
-#include "domain/value_data/types.h"
-#include "utils/tagged.h"
+#include "tagged.h"
+#include "value_data/types.h"
 
 namespace domain {
 
@@ -11,7 +12,7 @@ namespace detail {
 struct Profession {};
 }  // namespace detail
 
-using ProfessionId = utils::Tagged<int, detail::Profession>;
+using ProfessionId = utils::Tagged<uint64_t, detail::Profession>;
 using ProfessionIdHasher = utils::TaggedHasher<ProfessionId>;
 
 struct Profession {
