+++ src/domain/entities/shop/staff_position/staff_position.h
@@ -1,22 +1,23 @@
 #pragma once
 
-#include "utils/tagged.h"
-#include "domain/value_data/types.h"
-
+#include <cstdint>
 #include <optional>
+
+#include "tagged.h"
+#include "value_data/types.h"
 
 namespace domain {
 
 namespace detail {
-    struct StaffPosition{};
-} // namespace detail
+struct StaffPosition {};
+}  // namespace detail
 
-using StaffPositionId = utils::Tagged<int, detail::StaffPosition>;
+using StaffPositionId = utils::Tagged<uint64_t, detail::StaffPosition>;
 using StaffPositionIdHasher = utils::TaggedHasher<StaffPositionId>;
 
 struct StaffPosition {
     StaffPositionId staff_position_id;
     std::string name;
 };
-    
-} // namespace domain
+
+}  // namespace domain
