+++ src/domain/entities/timesheet/admin_categorys/admin_category.h
@@ -1,17 +1,18 @@
 #pragma once
 
-#include "utils/tagged.h"
-#include "domain/value_data/types.h"
-
+#include <cstdint>
 #include <string>
+
+#include "tagged.h"
+#include "value_data/types.h"
 
 namespace domain {
 
 namespace detail {
-    struct AdminCategoryIdTag{};
-} // namespace detail
+struct AdminCategoryIdTag {};
+}  // namespace detail
 
-using AdminCategoryId = utils::Tagged<int, detail::AdminCategoryIdTag>;
+using AdminCategoryId = utils::Tagged<uint64_t, detail::AdminCategoryIdTag>;
 using AdminCategoryIdHasher = utils::TaggedHasher<AdminCategoryId>;
-    
-} // namespace domain
+
+}  // namespace domain
