+++ src/domain/entities/timesheet/holidays/holiday.h
@@ -1,19 +1,20 @@
 #pragma once
 
-#include "utils/tagged.h"
-#include "domain/value_data/types.h"
-
+#include <cstdint>
 #include <string>
+
+#include "tagged.h"
+#include "value_data/types.h"
 
 namespace domain {
 
 namespace detail {
-    struct HolidayIdtag{};
-    struct ExtraHolidayDayIdTag{};
-    struct PreHolidayDayIdTag{};
-} // namespace detail
+struct HolidayIdtag {};
+struct ExtraHolidayDayIdTag {};
+struct PreHolidayDayIdTag {};
+}  // namespace detail
 
-using HolidayId = utils::Tagged<int, detail::HolidayIdtag>;
+using HolidayId = utils::Tagged<uint64_t, detail::HolidayIdtag>;
 using HolidayIdHasher = utils::TaggedHasher<HolidayId>;
-   
-} // namespace domain
+
+}  // namespace domain
