+++ src/domain/entities/user/user.h
@@ -1,13 +1,14 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <unordered_map>
 
-#include "domain/entities/permission/permission.h"
-#include "domain/entities/shop/department/department.h"
-#include "domain/entities/shop/employee/employee.h"
-#include "domain/entities/timesheet/admin_categorys/admin_category.h"
-#include "utils/tagged.h"
+#include "entities/permission/permission.h"
+#include "entities/shop/department/department.h"
+#include "entities/shop/employee/employee.h"
+#include "entities/timesheet/admin_categorys/admin_category.h"
+#include "tagged.h"
 
 namespace domain {
 
@@ -15,7 +16,7 @@ namespace detail {
 struct UserIdTag {};
 }  // namespace detail
 
-using UserId = utils::Tagged<int, detail::UserIdTag>;
+using UserId = utils::Tagged<uint64_t, detail::UserIdTag>;
 using UserIdHasher = utils::TaggedHasher<UserId>;
 using UserPermissions = std::unordered_map<DepartmentId, Permissions, DepartmentIdHasher>;
