+++ src/infrastructure/repositories_impl/permission_repository/permission_repository.cpp
@@ -1,10 +1,12 @@
-#include "permission_repository.h"
+#include "repositories_impl/permission_repository/permission_repository.h"
 
-#include "domain/entities/permission/permission.h"
-#include "domain/entities/shop/department/department.h"
-#include "domain/entities/user/user.h"
-#include "infrastructure/connection_pool/querys.h"
-#include "infrastructure/json_formater/json_formater.h"
+#include <cstdint>
+
+#include "connection_pool/querys.h"
+#include "entities/permission/permission.h"
+#include "entities/shop/department/department.h"
+#include "entities/user/user.h"
+#include "json_formater/json_formater.h"
 
 using namespace infrastructure;
 
@@ -18,10 +20,10 @@ std::optional<domain::UserPermissions> PermissionRepository::downloadUserPermiss
     domain::UserPermissions user_permissions;
 
     for (const auto row : result) {
-        auto db_deprtment_id = row[tables::permissions::DEPARTMENT_ID].as<int>();
+        auto db_deprtment_id = row[tables::permissions::DEPARTMENT_ID].as<uint64_t>();
         domain::DepartmentId department_id{db_deprtment_id};
 
-        auto db_permissions = row[tables::permissions::PERMISSIONS].as<int>();
+        auto db_permissions = row[tables::permissions::PERMISSIONS].as<uint64_t>();
         domain::Permissions permissions{static_cast<unsigned long>(db_permissions)};
 
         user_permissions.emplace(department_id, permissions);
