+++ src/infrastructure/token_manager/token_manager.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <jwt-cpp/jwt.h>
+#include <cstdint>
 #include <optional>
 #include <string_view>
 #include <unordered_map>
@@ -9,10 +9,10 @@ namespace infrastructure {
 
 class TokenManager {
 public:
-    using Payload = std::optional<std::unordered_map<std::string, int>>;
+    using Payload = std::optional<std::unordered_map<std::string, uint64_t>>;
     using Token = std::optional<std::string>;
 
-    Token createToken(int user_id, int employee_id) const;
+    Token createToken(uint64_t user_id, uint64_t employee_id) const;
     Payload getPayloadFromToken(std::string_view token) const;
 };
